project3_freezing_strategy/pyproject.toml
[project]
name = "indiwaste-autoresearch-project3"
version = "0.1.0"
description = "Karpathy-style autoresearch loop for IndiWASTE Project 3: transfer learning freezing strategies"
readme = "program.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "pandas>=2.3",
    "pillow>=10.0",
    "scikit-learn>=1.5",
    "torch>=2.9",
    "torchvision>=0.24",
]

[tool.uv]
package = false
